backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0

# File Upload Handling
python-multipart==0.0.20

# LangChain and AI Model
langchain==0.3.13
langchain-google-genai==2.0.8

# Environment Variables
python-dotenv==1.0.1

# File Parsing Libraries
PyPDF2==3.0.1
python-docx==1.1.2

# Data Validation
pydantic==2.10.5